runtime/consensus/tendermint/verifier: Split verifier into modules
